Convert to go modules, drop use of testify
This commits drops the direct dependency on testify, and adds a go.mod
and go.sum file.

Some of the longer tests are stuctured and named according to GWT
(Given-When-Then) principals, by using normal Go sub-tests. It however
dropps the outer precondition level (Given), as there was no strong need
for it for it in the tests in question.

- https://en.wikipedia.org/wiki/Given-When-Then
- https://www.agilealliance.org/glossary/gwt
- https://en.wikipedia.org/wiki/Cucumber_(software)#Syntax